requirements.txt
GeoAlchemy2==0.14.0
geopandas==0.13.2
h3==3.7.7
numpy==1.24.3
p_tqdm==1.4.0
pandas==2.2.2
PyYAML==6.0
Requests==2.32.3
rpy2==3.5.11
scikit_learn==1.4.1.post1
scipy==1.10.1
Shapely==2.0.5
SQLAlchemy==2.0.18
statsmodels==0.14.0
timezonefinder==6.5.2
tqdm==4.65.0
utm==0.7.0
wquantiles==0.6
